refactored backlinking
